refactor: split incremental model types
Amp-Thread-ID: https://ampcode.com/threads/T-019ee101-af22-75ba-b281-81391a483184
